Progressbar: Move progressbar into widgets folder
