+++ tests/unit/progressbar/common.js
@@ -1,6 +1,6 @@
 define( [
 	"lib/common",
-	"ui/progressbar"
+	"ui/widgets/progressbar"
 ], function( common ) {
 
 common.testWidget( "progressbar", {
